FrontImage/requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
boto3==1.28.12
botocore==1.31.12
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.10.0
einops==0.2.0
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
fire==0.5.0
fonttools==4.41.1
fsspec==2023.6.0
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h11==0.14.0
huggingface-hub==0.16.4
idna==2.10
imageio==2.31.1
imgaug==0.4.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
lazy_loader==0.3
lmdb==1.4.1
mangum==0.17.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow
prefetch-generator==1.0.1
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydantic==2.1.1
pydantic_core==2.4.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.2
rsa==4.9
s3transfer==0.6.1
sahi==0.11.14
scikit-image
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
supervision==0.12.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.7.18
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.143
urllib3==1.26.16
uvicorn==0.23.1
vietocr==0.3.11
Werkzeug==2.3.6
wget==3.2
yolov5==7.0.12
